feat: refactor concurrency test for A2A clients with variable transport
