--- infra/docker/configs/mosquitto/mosquitto.conf
@@ -1,5 +1,0 @@
-listener 1883
-allow_anonymous true
-
-listener 9001
-protocol websockets
