+++ infra/docker/docker-compose.yaml
@@ -33,7 +33,7 @@ services:
     image: clickhouse/clickhouse-server
     container_name: clickhouse-server
     ports:
-      - "9000:9000"
+      - "9004:9000"
       - "8123:8123"
     environment:
       CLICKHOUSE_USER: admin
@@ -65,7 +65,7 @@ services:
       - "4317:4317"
       - "4318:4318"
     volumes:
-      - ./config/docker/otel/otel-collector-config.yaml:/etc/otel-collector-config.yaml
+      - ./configs/otelcollector/otel-collector-config.yaml:/etc/otel-collector-config.yaml
     command: ["--config", "/etc/otel-collector-config.yaml"]
     depends_on:
       clickhouse-server:
